#165 - Enable the copy of nodes that do not match the internal data structure (children)
Components are copied by first creating a structure based on their component and then the properties are overlaid. For containers that does not work as they work with different components. Here we look for the children’s component first by using the path to find the source node for its type and then creating them based on their component before overlaying the properties.